Initial Docker run is working (see details)
Needed to update the host to be `miedb` and also need to add data to the Docker container db with commands like the below. Still not sure why this is necessary and starting a docker container doesn't just start a container with my existing db data?

docker exec -it miedb mysql --user=app --password=wonderful miechallenge -e "CREATE TABLE Games (
    game_id INT NOT NULL AUTO_INCREMENT,
    game_name VARCHAR(255) NOT NULL,
    game_image LONGBLOB,
    PRIMARY KEY (game_id)
);"